pyproject.toml
[tool.poetry]
name = "georgebot"
version = "0.5.0"
description = "Simple Deribit trading bot"
authors = ["me"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
asyncio = "^3.4.3"
websockets = "^8.1"
PyYAML = "^5.4.1"
peewee = "^3.14.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
